// client/src/lib/Types.ts
export type FieldType = {
  title: string;
  description?: string;
  goals: string;
  collectionSkills: string[];
  totalTime: string;
  resources: string;
  stepByStep: string;
};

export type Task = {
  _id: string;
  title: string;
  description?: string;
  goals: string;
  collectionSkills: string[];
  totalTime: string;
  resources: string;
  stepByStep: string;
};